requirements_dev.txt
pytest==6.2.2
pytest-django==4.1.0
factory-boy==3.2.0
